test: move tests 'multiple inherited' to seperate file
